+++ src/r_interface.c
@@ -225,6 +225,24 @@ extern SEXP R_ggml_set_op_params_i32(SEXP, SEXP, SEXP);
 extern SEXP R_ggml_get_op_params_f32(SEXP, SEXP);
 extern SEXP R_ggml_set_op_params_f32(SEXP, SEXP, SEXP);
 
+// Timestep embedding (defined in r_interface_graph.c)
+extern SEXP R_ggml_timestep_embedding(SEXP, SEXP, SEXP, SEXP);
+
+// CPU-side tensor data access (defined in r_interface_graph.c)
+extern SEXP R_ggml_set_f32_nd(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_ggml_get_f32_nd(SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_ggml_get_i32_nd(SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_ggml_set_i32_nd(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_ggml_tensor_nb(SEXP);
+extern SEXP R_ggml_backend_tensor_get_and_sync(SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_ggml_backend_tensor_get_f32(SEXP);
+extern SEXP R_ggml_tensor_num(SEXP);
+extern SEXP R_ggml_tensor_data_ptr(SEXP);
+extern SEXP R_ggml_tensor_copy(SEXP, SEXP);
+extern SEXP R_ggml_tensor_set_f32_scalar(SEXP, SEXP);
+extern SEXP R_ggml_get_first_tensor(SEXP);
+extern SEXP R_ggml_get_next_tensor(SEXP, SEXP);
+
 // Low-level quantization functions (defined in r_interface_quants.c)
 // Dequantize row functions
 extern SEXP R_dequantize_row_q4_0(SEXP, SEXP);
@@ -1738,6 +1756,24 @@ static const R_CallMethodDef CallEntries[] = {
 
     // Quantization info
     {"R_ggml_quant_block_info",                 (DL_FUNC) &R_ggml_quant_block_info,                  1},
+
+    // Timestep embedding
+    {"R_ggml_timestep_embedding",               (DL_FUNC) &R_ggml_timestep_embedding,                4},
+
+    // CPU-side tensor data access
+    {"R_ggml_set_f32_nd",                       (DL_FUNC) &R_ggml_set_f32_nd,                        6},
+    {"R_ggml_get_f32_nd",                       (DL_FUNC) &R_ggml_get_f32_nd,                        5},
+    {"R_ggml_get_i32_nd",                       (DL_FUNC) &R_ggml_get_i32_nd,                        5},
+    {"R_ggml_set_i32_nd",                       (DL_FUNC) &R_ggml_set_i32_nd,                        6},
+    {"R_ggml_tensor_nb",                        (DL_FUNC) &R_ggml_tensor_nb,                         1},
+    {"R_ggml_backend_tensor_get_and_sync",      (DL_FUNC) &R_ggml_backend_tensor_get_and_sync,       4},
+    {"R_ggml_backend_tensor_get_f32",           (DL_FUNC) &R_ggml_backend_tensor_get_f32,            1},
+    {"R_ggml_tensor_num",                       (DL_FUNC) &R_ggml_tensor_num,                        1},
+    {"R_ggml_tensor_data_ptr",                  (DL_FUNC) &R_ggml_tensor_data_ptr,                   1},
+    {"R_ggml_tensor_copy",                      (DL_FUNC) &R_ggml_tensor_copy,                       2},
+    {"R_ggml_tensor_set_f32_scalar",            (DL_FUNC) &R_ggml_tensor_set_f32_scalar,             2},
+    {"R_ggml_get_first_tensor",                 (DL_FUNC) &R_ggml_get_first_tensor,                  1},
+    {"R_ggml_get_next_tensor",                  (DL_FUNC) &R_ggml_get_next_tensor,                   2},
 
     {NULL, NULL, 0}
 };